Deployment: .Net Client now used, not full framework
Projects re-targeted to just use the Client framework and not the full
framework. This will increase the likelihood of the user already having
it installed.
Dapper reinstalled to use the revised framework
SqlCompact ref ref changed to copy local so that it gets included in the
private deployment file set